Moving Sequel Dirty plugin code to own module
